+++ frontends/sdl/src/util.rs
@@ -3,7 +3,7 @@ use sdl2::{
     AudioSubsystem, EventPump, TimerSubsystem, VideoSubsystem,
 };
 
-/// Structure that provide the complete set of Graphics
+/// Structure that provides the complete set of Graphics
 /// and Sound syb-system ready to be used by the overall
 /// emulator infrastructure.
 pub struct Graphics {
@@ -44,7 +44,12 @@ impl Graphics {
 
         // creates an accelerated canvas to be used in the drawing
         // then clears it so that is can be presented empty initially
-        let mut canvas = window.into_canvas().accelerated().build().unwrap();
+        let mut canvas = window
+            .into_canvas()
+            .accelerated()
+            .present_vsync()
+            .build()
+            .unwrap();
         canvas.set_logical_size(width, height).unwrap();
         canvas.clear();
 
@@ -67,6 +72,8 @@ impl Graphics {
     }
 }
 
+/// Creates an SDL2 Surface structure from the provided
+/// bytes that represent an image (eg: an PNG image buffer).
 pub fn surface_from_bytes(bytes: &[u8]) -> Surface {
     unsafe {
         let rw_ops = RWops::from_bytes(bytes).unwrap();